requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
asgiref==3.11.0
atpublic==5.1
attrs==25.4.0
bigframes==2.31.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
db-dtypes==1.5.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gcsfs==2025.12.0
geopandas==1.1.1
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-bigquery==3.39.0
google-cloud-bigquery-connection==1.19.0
google-cloud-bigquery-storage==2.35.0
google-cloud-core==2.5.0
google-cloud-functions==1.21.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.7.0
google-cloud-storage-control==1.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
htmltools==0.6.0
humanize==4.14.0
idna==3.11
ipywidgets==8.1.8
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
linkify-it-py==2.0.3
markdown-it-py==4.0.0
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
multidict==6.7.0
narwhals==2.14.0
numpy==2.3.5
oauthlib==3.3.1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pandas-gbq==0.32.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
questionary==2.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.2
shiny==1.5.1
shinychat==0.2.8
shinyswatch==0.9.0
six==1.17.0
sqlglot==28.4.1
starlette==0.50.0
tabulate==0.9.0
toolz==1.1.0
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
urllib3==2.6.2
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
widgetsnbextension==4.0.15
yarl==1.22.0
